Move metrics to analyzer and remove unsused packages
