+++ pkg/analyzer/metrics.go
@@ -1,4 +1,4 @@
-package controller
+package analyzer
 
 import (
 	"fmt"
