--- pkg/config/defaults.go
@@ -1,5 +1,0 @@
-package config
-
-var (
-	HttpDefaultHost = "http://localhost:5050"
-)
